Controller polishing
This change cleans up the controllers a bit.  The biggest change was
creating an OpenMatchFinder support class.  This class is not a
controller, not repository, and not a domain model, and therefore is
unique within the system.  It demonstrates how Spring can both wire
and use for wiring, arbitrary classes within an application.